Made name and friend count come from the viewmodel
